varnishd: add lock duration
Our instrumented locks lacks a way to tell how long a lock was held
before released. Just looking at the number of time the lock was taken
is not enough. Some locks are more expensive than others (held for
longer).

This commit adds duration to each instrumented lock, the resolution
depend on the parameter 'tick_rate'.

Signed-off-by: Asad Sajjad Ahmed <asadsa@varnish-software.com>