backend: drop two of the helper functions
Now that the connection queue is abstracted into its own function and the
operation to dequeue and finalize can only be done by this function it
is time to simplify things by removing some of the helper functions.

We maintain the signal helper function as this happens async to us
waiting for connections.

Signed-off-by: Asad Sajjad Ahmed <asadsa@varnish-software.com>